adding and using laravel spatie packages

+++ app/Http/Controllers/Manage/Salary/SalaryController.php
@@ -14,8 +14,15 @@ use PDF;
 
 class SalaryController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware(['permission:create salary|read salary|update salary|delete salary|print salary']);
+    }
+
     public function index()
     {
+        $this->authorize('read salary');
+
         return view('manage.salary.index', [
             'salaries' => Salary::orderBy('month')->get()
         ]);
@@ -23,6 +30,8 @@ class SalaryController extends Controller
 
     public function show(Salary $salary)
     {
+        $this->authorize('read salary');
+
         return view('manage.salary.show', [
             'salary' => $salary
         ]);
@@ -30,9 +39,9 @@ class SalaryController extends Controller
 
     public function create()
     {
-        $users = User::whereHas('role', function ($q) {
-            $q->whereIn('name', ['staff', 'teacher']);
-        })->orderBy('name')->get();
+        $this->authorize('create salary');
+
+        $users = User::role(['staff', 'teacher'])->orderBy('name')->get();
 
         return view('manage.salary.create', [
             'users' => $users,
@@ -43,8 +52,10 @@ class SalaryController extends Controller
 
     public function store(Request $request)
     {
+        $this->authorize('create salary');
+
         $request->validate([
-            'month' => 'required|date',
+            'month' => 'required|date|unique:salaries,month',
             'status' => 'required'
         ]);
 
@@ -94,8 +105,74 @@ class SalaryController extends Controller
         return redirect(route('app.salaries.index'))->with('success', 'Penggajian Berhasil Dibuat');
     }
 
+    public function edit(Salary $salary)
+    {
+        $this->authorize('update salary');
+
+        $users = User::role(['staff', 'teacher'])->orderBy('name')->get();
+
+        return view('manage.salary.edit', [
+            'users' => $users,
+            'salary' => $salary,
+            'salaryDetails' => SalaryDetail::where('salary_id', $salary->id)->get(),
+            'allowances' => Allowance::all(),
+            'salary_cuts' => SalaryCut::all()
+        ]);
+    }
+
+    public function update(Salary $salary, Request $request)
+    {
+        $this->authorize('update salary');
+
+        $request->validate([
+            'month' => 'required|date|unique:salaries,month,' . $salary->id,
+            'status' => 'required'
+        ]);
+
+        if (!$request->users) {
+            return back()->with('error', 'Harap mencentang Staff / Guru');
+        }
+
+        $update = [];
+        $components = [];
+
+        foreach ($request->users as $indexUser => $user) {
+            // Allowance
+            foreach ($request->allowances[$user] as $allowanceId => $allowance) {
+                $_allowance = Allowance::find($allowanceId);
+
+                $components[$user]['allowances'][] = [
+                    'allowance_id' => $allowanceId,
+                    'name' => $_allowance->name,
+                    'amount' => (int) cleanCurrency($allowance)
+                ];
+            }
+            // Salary Cuts
+            foreach ($request->salary_cuts[$user] as $salaryCutsId => $salary_cut) {
+                $_salary_cut = SalaryCut::find($salaryCutsId);
+
+                $components[$user]['salary_cuts'][] = [
+                    'salary_cut_id' => $salaryCutsId,
+                    'name' => $_salary_cut->name,
+                    'amount' => (int) cleanCurrency($salary_cut)
+                ];
+            }
+
+            $update[] = [
+                'uid' => Str::uuid(),
+                'components' => json_encode($components[$user]),
+                'user_id' => $user,
+                'salary_id' => $salary->id,
+                'created_at' => date('Y-m-d H:i:s'),
+                'updated_at' => date('Y-m-d H:i:s')
+            ];
+        }
+    }
+
     public function destroy(Salary $salary)
     {
+        $this->authorize('delete salary');
+
         $salary->details()->delete();
         $salary->delete();
 
@@ -104,6 +181,8 @@ class SalaryController extends Controller
 
     public function generatePDF(SalaryDetail $salary_detail, $type)
     {
+        $this->authorize('print salary');
+
         $detail = SalaryDetail::where('id', $salary_detail->id)->with('user')->first();
 
         $data = [
@@ -111,7 +190,7 @@ class SalaryController extends Controller
             'detail' => $detail->toArray(),
             'components' => json_decode($detail->components, true),
             'user' => $detail->user->toArray(),
-            'role' => $detail->user->role->toArray(),
+            'role' => $detail->user->roles->first()->toArray(),
             'last_education' => $detail->user->last_education->toArray(),
             'positions' => $detail->user->positions->toArray(),
         ];
